Add dinamic routes

+++ src/Services/user_services.py
@@ -1,15 +1,28 @@
 from Services.file_services import create_file, delete_folder, read_json
+from helpers.get_route import get_route
 from typing import Dict
 
-def create_user(userData: Dict[str, str], folder_path: str ):
+def create_user(userData: Dict[str, str], folder_path: str, admin_password: str ):
+    admin = _get_admin()
+    if admin_password != admin["password"]:
+        raise ValueError("Invalid credential!, admin password wrong")
+
     create_file(userData, folder_path , f"{userData["code"]}.json")
 
-def delete_user(user_code: str):
-    folder_path = f"C:\\Users\\mon_e\\OneDrive\\Escritorio\\Proyectos\\UNIVERSIDAD\\assists_tracker\\Data\\{user_code}"
+def delete_user(user_code: str, admin_password: str):
+    admin = _get_admin()
+    if admin_password != admin["password"]:
+        raise ValueError("Invalid credential!, admin password wrong")
+
+    folder_path = get_route(f"Data/{user_code}")
     delete_folder(folder_path)
 
-def edit_user(user_code: str, new_user_data: Dict[str, str]):
-    file_path = f"C:\\Users\\mon_e\\OneDrive\\Escritorio\\Proyectos\\UNIVERSIDAD\\assists_tracker\\Data\\{user_code}"
+def edit_user(user_code: str, new_user_data: Dict[str, str], admin_password: str):
+    admin = _get_admin()
+    if admin_password != admin["password"]:
+        raise ValueError("Invalid credential!, admin password wrong")
+
+    file_path = get_route(f"Data/{user_code}")
     user_saved = read_json(f"{file_path}\\{user_code}.json")
     user_updated = {
         **user_saved,
@@ -19,6 +32,11 @@ def edit_user(user_code: str, new_user_data: Dict[str, str]):
     create_file(user_updated, f"{file_path}", f"{user_code}.json")
         
 def get_user(user_code: str):
-    file_path = f"C:\\Users\\mon_e\\OneDrive\\Escritorio\\Proyectos\\UNIVERSIDAD\\assists_tracker\\Data\\{user_code}"
+    file_path = get_route(f"Data/{user_code}")
     user_saved = read_json(f"{file_path}\\{user_code}.json")
     return user_saved
+
+def _get_admin():
+    file_path = get_route(f"Data/admin/admin.json")
+    admin = read_json(file_path)
+    return admin
